gather root/super/evm/keccak circuits tests
