+++ zkevm-circuits/src/bytecode_circuit.rs
@@ -6,5 +6,5 @@ pub mod bytecode_unroller;
 pub mod circuit;
 pub(crate) mod param;
 /// Bytecode circuit tester
-#[cfg(any(feature = "test", test))]
+#[cfg(any(feature = "test", test, feature = "test-circuits"))]
 pub mod test;
