+++ zkevm-circuits/src/super_circuit/test.rs
@@ -1,0 +1,184 @@
+use super::*;
+use ethers_signers::{LocalWallet, Signer};
+use halo2_proofs::dev::MockProver;
+use halo2_proofs::halo2curves::bn256::Fr;
+use log::error;
+use mock::{TestContext, MOCK_CHAIN_ID};
+use rand::SeedableRng;
+use rand_chacha::ChaCha20Rng;
+use std::collections::HashMap;
+
+use eth_types::{address, bytecode, geth_types::GethData, Word};
+
+#[test]
+fn super_circuit_degree() {
+    let mut cs = ConstraintSystem::<Fr>::default();
+    SuperCircuit::<_, 1, 32, 0x100>::configure(&mut cs);
+    log::info!("super circuit degree: {}", cs.degree());
+    log::info!("super circuit minimum_rows: {}", cs.minimum_rows());
+    assert!(cs.degree() <= 9);
+}
+
+fn test_super_circuit<
+    const MAX_TXS: usize,
+    const MAX_CALLDATA: usize,
+    const MOCK_RANDOMNESS: u64,
+>(
+    block: GethData,
+    circuits_params: CircuitsParams,
+) {
+    let (k, circuit, instance, _) =
+        SuperCircuit::<Fr, MAX_TXS, MAX_CALLDATA, MOCK_RANDOMNESS>::build(block, circuits_params)
+            .unwrap();
+    let prover = MockProver::run(k, &circuit, instance).unwrap();
+    let res = prover.verify_par();
+    if let Err(err) = res {
+        error!("Verification failures: {:#?}", err);
+        panic!("Failed verification");
+    }
+}
+
+pub(crate) fn block_1tx() -> GethData {
+    let mut rng = ChaCha20Rng::seed_from_u64(2);
+
+    let chain_id = (*MOCK_CHAIN_ID).as_u64();
+
+    let bytecode = bytecode! {
+        GAS
+        STOP
+    };
+
+    let wallet_a = LocalWallet::new(&mut rng).with_chain_id(chain_id);
+
+    let addr_a = wallet_a.address();
+    let addr_b = address!("0x000000000000000000000000000000000000BBBB");
+
+    let mut wallets = HashMap::new();
+    wallets.insert(wallet_a.address(), wallet_a);
+
+    let mut block: GethData = TestContext::<2, 1>::new(
+        None,
+        |accs| {
+            accs[0]
+                .address(addr_b)
+                .balance(Word::from(1u64 << 20))
+                .code(bytecode);
+            accs[1].address(addr_a).balance(Word::from(1u64 << 20));
+        },
+        |mut txs, accs| {
+            txs[0]
+                .from(accs[1].address)
+                .to(accs[0].address)
+                .gas(Word::from(1_000_000u64));
+        },
+        |block, _tx| block.number(0xcafeu64),
+    )
+    .unwrap()
+    .into();
+    block.sign(&wallets);
+    block
+}
+
+fn block_2tx() -> GethData {
+    let mut rng = ChaCha20Rng::seed_from_u64(2);
+
+    let chain_id = (*MOCK_CHAIN_ID).as_u64();
+
+    let bytecode = bytecode! {
+        GAS
+        STOP
+    };
+
+    let wallet_a = LocalWallet::new(&mut rng).with_chain_id(chain_id);
+
+    let addr_a = wallet_a.address();
+    let addr_b = address!("0x000000000000000000000000000000000000BBBB");
+
+    let mut wallets = HashMap::new();
+    wallets.insert(wallet_a.address(), wallet_a);
+
+    let mut block: GethData = TestContext::<2, 2>::new(
+        None,
+        |accs| {
+            accs[0]
+                .address(addr_b)
+                .balance(Word::from(1u64 << 20))
+                .code(bytecode);
+            accs[1].address(addr_a).balance(Word::from(1u64 << 20));
+        },
+        |mut txs, accs| {
+            txs[0]
+                .from(accs[1].address)
+                .to(accs[0].address)
+                .gas(Word::from(1_000_000u64));
+            txs[1]
+                .from(accs[1].address)
+                .to(accs[0].address)
+                .gas(Word::from(1_000_000u64));
+        },
+        |block, _tx| block.number(0xcafeu64),
+    )
+    .unwrap()
+    .into();
+    block.sign(&wallets);
+    block
+}
+
+const TEST_MOCK_RANDOMNESS: u64 = 0x100;
+
+// High memory usage test.  Run in serial with:
+// `cargo test [...] serial_ -- --ignored --test-threads 1`
+#[ignore]
+#[test]
+fn serial_test_super_circuit_1tx_1max_tx() {
+    let block = block_1tx();
+    const MAX_TXS: usize = 1;
+    const MAX_CALLDATA: usize = 32;
+    let circuits_params = CircuitsParams {
+        max_txs: MAX_TXS,
+        max_calldata: MAX_CALLDATA,
+        max_rws: 256,
+        max_copy_rows: 256,
+        max_exp_steps: 256,
+        max_bytecode: 512,
+        max_evm_rows: 0,
+        max_keccak_rows: 0,
+    };
+    test_super_circuit::<MAX_TXS, MAX_CALLDATA, TEST_MOCK_RANDOMNESS>(block, circuits_params);
+}
+#[ignore]
+#[test]
+fn serial_test_super_circuit_1tx_2max_tx() {
+    let block = block_1tx();
+    const MAX_TXS: usize = 2;
+    const MAX_CALLDATA: usize = 32;
+    let circuits_params = CircuitsParams {
+        max_txs: MAX_TXS,
+        max_calldata: MAX_CALLDATA,
+        max_rws: 256,
+        max_copy_rows: 256,
+        max_exp_steps: 256,
+        max_bytecode: 512,
+        max_evm_rows: 0,
+        max_keccak_rows: 0,
+    };
+    test_super_circuit::<MAX_TXS, MAX_CALLDATA, TEST_MOCK_RANDOMNESS>(block, circuits_params);
+}
+#[ignore]
+#[test]
+fn serial_test_super_circuit_2tx_2max_tx() {
+    let block = block_2tx();
+    const MAX_TXS: usize = 2;
+    const MAX_CALLDATA: usize = 32;
+    let circuits_params = CircuitsParams {
+        max_txs: MAX_TXS,
+        max_calldata: MAX_CALLDATA,
+        max_rws: 256,
+        max_copy_rows: 256,
+        max_exp_steps: 256,
+        max_bytecode: 512,
+        max_evm_rows: 0,
+        max_keccak_rows: 0,
+    };
+    test_super_circuit::<MAX_TXS, MAX_CALLDATA, TEST_MOCK_RANDOMNESS>(block, circuits_params);
+}
